reorg of boost small vector stuff + actually install the header
